Move stats_client into app_backend.clients.stats_client
