+++ app_backend/tests/test_episode_filtering.py
@@ -3,7 +3,7 @@ from typing import Dict
 import pytest
 from fastapi.testclient import TestClient
 
-from metta.app_backend.stats_client import StatsClient
+from metta.app_backend.clients.stats_client import StatsClient
 
 
 class TestEpisodeFiltering:
